pyproject.toml
[project]
name = "Diffulex"
version = "0.0.1"
authors = [
    { name = "Yijie Jin", email = "drewjin0827@gmail.com"},
]
maintainers = [
    { name = "DENG Lab @ SJTU" }
]
description = "Flexible Diffusion LLM serving engine supporting multiple variants of dLLM decoding strategies with high efficiency and low latency."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "accelerate>=1.9.0",
    "bpytop>=1.0.68",
    "datasets>=4.0.0",
    "easydict>=1.13",
    "einops>=0.8.1",
    "keep-gpu>=0.2.0",
    "numpy>=2.3.1",
    "nvitop>=1.5.2",
    "peft>=0.16.0",
    "pip>=25.1.1",
    "pytest>=8.4.1",
    "pytest-forked>=1.6.0",
    "pyyaml>=6.0.2",
    "setuptools>=80.9.0",
    "torch",
    "tqdm>=4.67.1",
    "transformers>=4.53.2",
    "viztracer>=1.0.4",
    "matplotlib>=3.10.5",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pyzmq>=26.0.0",
    "msgpack>=1.0.0",
    "pandas>=2.3.3",
    "rich>=13.0.0",
    "colorama>=0.4.6",
    "lm-eval"
]

[project.urls]
Homepage = "https://sjtu-deng-lab.github.io/Diffulex"
Repository = "https://github.com/SJTU-DENG-Lab/Diffulex"
Organization = "https://github.com/SJTU-DENG-Lab"

[tool.ruff]
line-length = 120  # Avoid excessive line breaks (default 88 too aggressive)

[tool.ruff.lint.pycodestyle]
max-line-length = 120  # Align with formatter

[tool.pytest.ini_options]
markers = [
    "diffulex_dry_run: Diffulex engine dry-run tests (GPU, spawns TP workers; run with --forked)",
]

[tool.setuptools.packages.find]
include = [
    "diffulex",
    "diffulex_bench",
    "diffulex_kernel",
    "test"
]
